// web-booking-js-html/front/src/enums.ts
export enum SortDirectionEnum {
    ascend = "ASC",
    descend = "DESC"
}

export enum CitiesEnum {
    moscow = "Москва",
    chelyabinsk = "Челябинск",
    ekaterinburg = "Екатеринбург",
    krasnodar = "Краснодар",
    perm = "Пермь",
    novosibirsk = "Новосибирск",
    samara = "Самара",
    ufa = "Уфа"
}

export enum SortTypeEnum {
    title = "TITLE",
    price = "PRICE"    
}

export enum FiltersEnum {
    genre = "genre",
    ageContent = "age",
    price = "price"
}

export enum GenreEnum {
    comics = "Комиксы",
    romance = "Романтика",
    detective = "Детектив",
    action = "Боевик"
}

export enum AgeContentEnum {
    six = "6",
    twelve = "12",
    eighteen = "18"
}

export enum PriceEnum {
    min = "min",
    max = "max"
}

export enum ButtonTypeEnum {
    buy = "buy",
    corrent = "correct"
}
  